sql relationship update

+++ fastapi1/app/routers/post.py
@@ -11,12 +11,12 @@ router = APIRouter(
 
 
 @router.get("/",response_model=List[schemas.Post])
-def get_posts(db: Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
+def get_posts(db: Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user),limit: id = 10):
     # cursor.execute("""SELECT * FROM posts """)
     # posts = cursor.fetchall()
-    posts = db.query(models.Post).all()
-    return  posts
-
+    print(limit)
+    posts = db.query(models.Post).limit(limit).all()
+    return posts
 
 @router.post("/", status_code=status.HTTP_201_CREATED , response_model=schemas.Post)
 def create_posts(post: schemas.PostCreate,db: Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
@@ -27,8 +27,8 @@ def create_posts(post: schemas.PostCreate,db: Session = Depends(get_db),current_
     # )
     # new_post = cursor.fetchone()
     # conn.commit()
-    print(current_user.email)
-    new_post=models.Post(**post.dict()) #or models.Post(title=post.title,content=post.content,published=post.published)
+    print(current_user.id)
+    new_post=models.Post(owner_id = current_user.id,**post.dict()) #or models.Post(title=post.title,content=post.content,published=post.published)
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -39,7 +39,7 @@ def create_posts(post: schemas.PostCreate,db: Session = Depends(get_db),current_
 def get_post(id: int,db: Session=Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
     # cursor.execute(""" select * from posts where id = %s """, (str(id),))
     # post = cursor.fetchone()
-    post = db.query(models.Post).filter(models.Post.id== id).first()
+    post = db.query(models.Post).filter(models.Post.id == id).first()
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
@@ -52,12 +52,19 @@ def get_post(id: int,db: Session=Depends(get_db),current_user: int = Depends(oau
 def delete_post(id: int,db:Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
     # cursor.execute(""" DELETE FROM posts WHERE id = %s RETURNING  * """, (str(id)))
     # post = cursor.fetchone()
+
+    
     post = db.query(models.Post).filter(models.Post.id == id).first()
+    
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Posts of id:{id} is not present in the database",
         )
+    
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail=f"Not authorized to perform requested action")
+    
     db.delete(post)
     db.commit()
     # conn.commit()
@@ -74,11 +81,16 @@ def update_post(id: int, postt: schemas.PostCreate,db:Session = Depends(get_db),
     # conn.commit()
     post_query = db.query(models.Post).filter(models.Post.id == id)
     post = post_query.first()
+
     if post == None:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"post with id:{id} does not exist",
         )
+    
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail=f"Not authorized to perform requested action")
+    
     post_query.update(postt.dict(),synchronize_session=False)
     db.commit()
     # post_query.update('title':'hey this is my update title')
